test: extract shared redeploy-restart-flag helpers (address review comment)
authHeader/operation/getRestartRequired/readVersion were byte-for-byte duplicated
across redeploy-restart-flag.test.ts and redeploy-restart-flag-deletion.test.ts.
Extracted into redeploy-restart-flag-helpers.ts so a future fix to one (e.g. the
ECONNREFUSED tolerance in readVersion) only needs applying once.

Verified: both integration tests pass unchanged, build/lint/format clean.
